imperfect implementation of dumb ai
if you choose single player mode, the ai only works in the sense
that it automatically makes a move after the player's turn. BUT,
because of improper Player object initializations, it unfortunately
moves the RED pieces instead of white. I think this might be an issue
due to "this" in Player class, especially `get_pieces` cause that is
the method responsible for getting the correct pieces based on color
info of player. I do not see the mistake, but then its 3:30 in the morning
and I am fucking tired.

Comments have been added to the AI class on how it has been implemented.

I have not written code code nor did I make good classes. I just wanted
to get it done or working somehow.

**Sidenotes**:
1. For some strange reason, in double player, the first move
by red piece as [b6]X[a5] does not work - it treats it as an invalid
move - is it invalid actually? If not then getAllMoves method is faulty
and needs to be corrected.
2. Why do we have static everywhere? If you use static everywhere
then essentailly you are not really using OO concepts. It does not
look right to me. I have removed it where ever it was not needed in
my implementaitons so far, but I have not touched things that I did
not use so far. We may need to see where we are using static and
if we actually need it to be static.